requirements.txt
apiclient==1.0.4
asn1crypto==0.24.0
astroid==2.2.5
autopep8==1.4.4
cachetools==3.1.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
configparser==3.7.4
CouchDB==1.2
cryptography==2.7
feedparser==5.2.1
Flask==1.0.3
fritzconnection==0.6.5
future==0.17.1
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.13.0
idna==2.8
isort==4.3.18
kafka==1.3.5
lazy-object-proxy==1.4.0
lxml==4.3.3
mccabe==0.6.1
oauth2client==4.1.3
pyaes==1.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
pycparser==2.19
pylint==2.3.1
PySocks==1.7.0
python-dateutil==2.8.0
python-telegram-bot==12.0.0b1
requests==2.5.0
rsa==4.0
six==1.12.0
typed-ast==1.3.5
uritemplate==3.0.0
urllib3==1.25.3
wrapt==1.11.1
